CaviFinder is now separated from InSty
